requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.2.0
groq>=0.11.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pdfplumber>=0.11.0
unstructured>=0.15.0
streamlit>=1.39.0
python-dotenv>=1.0.0
pydantic>=2.0.0
yfinance>=0.2.0
edgar>=5.4.0
pytest>=8.0.0
deepeval>=1.0.0
tenacity>=9.0.0
